requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
prometheus-client==0.17.1
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.1.0
requests==2.31.0